// graphql/generated/fragments/calendar.fragment.ts
/* Do not edit this file. It was generated programmatically. */

import { Calendar } from "@/graphql/generated/models/calendar.model";
import { gql } from "@apollo/client";

export const calendarFragment = gql`
  fragment CalendarFragment on Calendar {
    __typename
    id
    userId
    name
    color
    provider
    remoteId
    syncToken
    accountId
    primary
    public
    enabled
    createdAt
    updatedAt
    archivedAt
  }
`;

export type CalendarFragment = Pick<
  Calendar,
  | "__typename"
  | "id"
  | "createdAt"
  | "updatedAt"
  | "archivedAt"
  | "userId"
  | "name"
  | "color"
  | "provider"
  | "remoteId"
  | "syncToken"
  | "accountId"
  | "primary"
  | "public"
  | "enabled"
>;